chore(jest): move jest tests in specific folders and exclude others tests from it
