+++ jest.config.js
@@ -3,4 +3,6 @@ export default {
   transform: {},
   testEnvironment: "node",
   moduleFileExtensions: ["js", "mjs", "cjs", "jsx", "ts", "tsx", "json", "node"],
+  roots: ["<rootDir>/tests/jest"],
+
 };
